backend/functions/get-ticket-status/requirements.txt
functions-framework==3.4.0
google-cloud-firestore==2.11.1